Merge integration with master at v0.8.1

+++ src/js/mutations/environment/RemovePackageComponentMutation.js
@@ -16,17 +16,20 @@ const mutation = graphql`
 
 function sharedUpdater(store, parentID, deletedId, connectionKey) {
   const environmentProxy = store.get(parentID);
-
-  const conn = RelayRuntime.ConnectionHandler.getConnection(
-    environmentProxy,
-    connectionKey,
-  );
-
-  if(conn){
-    RelayRuntime.ConnectionHandler.deleteNode(
-      conn,
-      deletedId,
+  if(environmentProxy) {
+    const conn = RelayRuntime.ConnectionHandler.getConnection(
+      environmentProxy,
+      connectionKey,
     );
+
+    if(conn){
+      RelayRuntime.ConnectionHandler.deleteNode(
+        conn,
+        deletedId,
+      );
+      store.delete(deletedId)
+    }
+    
   }
 }
